TextMetrics Web/App
TextMetrics is a React.js project designed to provide various text manipulation functionalities. The project includes features such as a word counter, character counter, removal of extra spaces, conversion to uppercase and lowercase, clearing the text, and the ability to copy the text.

The word counter functionality allows users to input text and instantly see the number of words present in the input. Similarly, the character counter displays the total number of characters in the text.

The removal of extra spaces feature helps in maintaining a clean and consistent text format. It identifies any unnecessary spaces between words or at the beginning or end of the text and removes them, ensuring a more readable output.

TextMetrics also enables users to convert the text to uppercase or lowercase, providing the flexibility to modify the case of the input as needed.

The "clear text" option allows users to quickly erase the content of the input field, providing a fresh start for new text entries.

In addition, the project includes a "copy text" functionality, which enables users to easily copy the text to their clipboard for further use or sharing.

Overall, TextMetrics is a React.js project that offers a range of useful text manipulation features, including word and character counting, removal of extra spaces, case conversion, text clearing, and copying text.